pyproject.toml
[project]
name = "tokay-artwork"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"cairosvg",
	"pillow",
]
